Add move-bitbucket-to-cloud.sh and templates for .config and .repositories
This script reads .config to use your data to connect to git server and
bitbucket.org. Then for each repository name in .repositories file it
clones the repo locally from your git server and creates repo with the
same name on bitbucket.org in you workspace/project, then it syncs the
repository to bitbucket.org.

It is assumed that all needed ssh keys are already setup beforehand.

Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>